recipe list adapter v0, next step is impl master detail layout and onclick

+++ app/src/main/java/com/example/bakingapp/model/IngredientModel.java
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName;
 
 public class IngredientModel {
     @SerializedName("quantity")
-    public Long quantity;
+    public Double quantity;
     @SerializedName("measure")
     public String measure;
     @SerializedName("ingredient")
